Move My Lists under a member profile page
/my-lists was a standalone, owner-only page. Moved it to /members/[username]
so members can also browse each other's public lists from the same place —
the natural next step now that custom lists are public, rather than only
being reachable via a direct /lists/[id] link someone already has.

Viewing your own profile shows Favorites, Watchlist, and full list
management (create/rename/delete), matching the old /my-lists exactly;
viewing someone else's shows only their public custom lists, read-only.
Favorites/Watchlist stay exactly as private as they've always been — never
shown to anyone but the signed-in owner, regardless of whose profile is
being viewed. /my-lists itself now just redirects to your own profile so
existing links keep working. Navbar's "My Lists" link and username display
both point at the new route.

Verified against a live dev server with three distinct signed-in states
(owner, a different signed-in member, anonymous) that Favorites/Watchlist
never appear except to the owner, while a public list appears identically
to all three; confirmed /my-lists redirects correctly for both signed-in
and anonymous visitors.
